add like watch func

+++ client/src/components/Item/index.jsx
@@ -4,9 +4,10 @@ import { connect } from 'react-redux'
 
 import {
     addItemToWishlist,
-    deleteItemFromWishlist
+    deleteItemFromWishlist,
+    updateWishlist,
+    getGenres
 } from '../../actions/itemsCollectionsActions'
-import { getGenres } from '../../actions/itemsCollectionsActions'
 
 import './index.css'
 
@@ -15,11 +16,20 @@ class Item extends Component {
         super(props);
 
         this.state = {
-            isInCollection: false
+            isInCollection: false,
+            wishlistItem: null
         }
 
         this.addToWishlist = this.addToWishlist.bind(this)
         this.deleteFromWishlist = this.deleteFromWishlist.bind(this)
+        this.itemAction = this.itemAction.bind(this)
+    }
+
+    itemAction(action) {
+        const {type, user} = this.props
+        const {wishlistItem} = this.state
+        this.props.updateWishlist(type, action, wishlistItem.id, user.userId, !wishlistItem[action])
+        // console.log(action, wishlistItem.id)
     }
 
     addToWishlist() {
@@ -56,11 +66,20 @@ class Item extends Component {
             return !1
         }
 
-        this.props.deleteItemFromWishlist(type, id, user.userId)
+        if (window.confirm("Delete item from wishlist?")) {
+            this.props.deleteItemFromWishlist(type, id, user.userId)
+        }
+    }
+
+    componentDidMount() {
+        const { user, type, id } = this.props
+        this.setState({
+            wishlistItem: user[type].find(item => item.id === id) || null
+        })
     }
 
     render() {
-        const { type, title, name, img, liked, watched, slug, id, userCollection, user, genre_ids } = this.props
+        const { type, title, name, img, slug, id, userCollection, user, genre_ids } = this.props
         let i = img || 'https://uoslab.com/images/tovary/no_image.jpg'
         let itemGenres = []
         if (genre_ids && this.props.moviesGenres) {
@@ -72,7 +91,10 @@ class Item extends Component {
         } else if (this.props.genres) {
             itemGenres = this.props.genres.map(genre => genre.name)
         }
+
+        const itemIds = user ? user[type].map(item => item.id) : []
         const searchItem = slug ? slug : id
+
         return (
 
             <div className={`movie_item ${type}`}>
@@ -92,20 +114,20 @@ class Item extends Component {
                     <div className="movie_info__bottom">
                         <p className="ml-auto">
                             {
-                                userCollection ?
+                                this.props.wishlist && this.state.wishlistItem ?
                                     <React.Fragment>
-                                        <span className="text-info pointer mx-2">
-                                            <i className={watched ? "fas fa-eye" : "far fa-eye"}></i>
+                                        <span className="text-info pointer mx-2" onClick={() => this.itemAction('watched')}>
+                                            <i className={this.state.wishlistItem.watched ? "fas fa-flag" : "far fa-flag"}></i>
                                         </span>
-                                        <span className="text-red pointer mx-2" title="like">
-                                            <i className={liked ? "fas fa-heart" : "far fa-heart"}></i>
+                                        <span className="text-red pointer mx-2" title="like" onClick={() => this.itemAction('liked')}>
+                                            <i className={this.state.wishlistItem.liked ? "fas fa-heart" : "far fa-heart"}></i>
                                         </span>
                                     </React.Fragment> : null
                             }
 
                             {
                                 user ?
-                                    user[type].includes(id) ?
+                                    itemIds.includes(id) ?
                                         <span className="ml-auto pointer" title="Remove from wishlist" onClick={this.deleteFromWishlist}><i className="fas fa-star text-warning"></i></span> :
                                         <span className="ml-auto pointer" title="Add to wishlist" onClick={this.addToWishlist}><i className="far fa-star"></i></span> : null
                             }
@@ -125,5 +147,6 @@ const mapStateToProps = state => ({
 export default connect(mapStateToProps, {
     addItemToWishlist,
     deleteItemFromWishlist,
+    updateWishlist,
     getGenres
 })(Item)
